requirements.txt
ipykernel
langchain
langchain-community
streamlit
python-dotenv 